+++ ratings-data-service/src/main/java/com/mielniczuk/recommendation/ratingsdataservice/repositories/RatingRepository.java
@@ -3,5 +3,5 @@ package com.mielniczuk.recommendation.ratingsdataservice.repositories;
 import com.mielniczuk.recommendation.ratingsdataservice.models.Rating;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RatingRepository extends JpaRepository<Rating, Integer> {
+public interface RatingRepository extends JpaRepository<Rating, Long> {
 }
