+++ ratings-data-service/src/main/java/com/mielniczuk/recommendation/ratingsdataservice/repositories/UserRatingRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRatingRepository extends JpaRepository<UserRating, Integer> {
+public interface UserRatingRepository extends JpaRepository<UserRating, Long> {
 }
